chore: move stylus specific chain into scaffold-stylus folder
