+++ packages/nextjs/app/blockexplorer/_components/SearchBar.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { isAddress, isHex } from "viem";
 import { usePublicClient } from "wagmi";
-import { arbitrumNitro } from "~~/utils/chain";
+import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
 
 export const SearchBar = () => {
   const [searchInput, setSearchInput] = useState("");
