+++ packages/nextjs/components/scaffold-eth/RainbowKitCustomConnectButton/BurnerWalletModal.tsx
@@ -2,7 +2,7 @@ import { useRef, useState } from "react";
 import { BlockieAvatar } from "..";
 import { useTheme } from "next-themes";
 import { useOutsideClick } from "~~/hooks/scaffold-eth/useOutsideClick";
-import { arbitrumNitro } from "~~/utils/chain";
+import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
 
 interface BurnerWalletModalProps {
   isOpen: boolean;
