+++ packages/nextjs/hooks/scaffold-eth/useScaffoldEventHistory.ts
@@ -6,7 +6,6 @@ import { Hash } from "viem";
 import { usePublicClient } from "wagmi";
 import { useDeployedContractInfo } from "~~/hooks/scaffold-eth";
 import scaffoldConfig from "~~/scaffold.config";
-import { arbitrumNitro } from "~~/utils/chain";
 import { replacer } from "~~/utils/scaffold-eth/common";
 import {
   ContractAbi,
@@ -14,6 +13,7 @@ import {
   UseScaffoldEventHistoryConfig,
   UseScaffoldEventHistoryData,
 } from "~~/utils/scaffold-eth/contract";
+import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
 
 /**
  * Reads events from a deployed contract
