+++ packages/nextjs/services/web3/wagmi-burner/burnerWalletConfig.ts
@@ -1,8 +1,8 @@
 import { Chain, Wallet } from "@rainbow-me/rainbowkit";
 import scaffoldConfig from "~~/scaffold.config";
 import { BurnerConnector, burnerWalletId, burnerWalletName } from "~~/services/web3/wagmi-burner/BurnerConnector";
-import { arbitrumNitro } from "~~/utils/chain";
 import { getTargetNetworks } from "~~/utils/scaffold-eth";
+import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
 
 const { onlyLocalBurnerWallet } = scaffoldConfig;
 const targetNetworks = getTargetNetworks();
