+++ packages/nextjs/utils/scaffold-eth/decodeTxData.ts
@@ -2,7 +2,7 @@ import { TransactionWithFunction } from "./block";
 import { GenericContractsDeclaration } from "./contract";
 import { Abi, decodeFunctionData, getAbiItem } from "viem";
 import contractData from "~~/contracts/deployedContracts";
-import { arbitrumNitro } from "~~/utils/chain";
+import { arbitrumNitro } from "~~/utils/scaffold-stylus/chain";
 
 type ContractsInterfaces = Record<string, Abi>;
 type TransactionType = TransactionWithFunction | null;
